safari-600.7-branch/LayoutTests/fast/dom/Range/range-insertNode-same-container-expected.txt

insertedElement = document.getElementById("div1");
range.setStart(containerElement, 3); range.setEnd(containerElement, 3); range.insertNode(insertedElement);
PASS range.startContainer is containerElement
PASS range.endContainer is containerElement
PASS range.startOffset is 2
PASS range.endOffset is 3
PASS range.toString() is "1"
range.setStart(containerElement, 3); range.setEnd(containerElement, 3); range.insertNode(insertedElement);
PASS range.startContainer is containerElement
PASS range.endContainer is containerElement
PASS range.startOffset is 2
PASS range.endOffset is 3
PASS range.toString() is "1"
insertedElement = document.getElementById("div3");
range.setStart(containerElement, 5); range.setEnd(containerElement, 5); range.insertNode(insertedElement);
PASS range.startContainer is containerElement
PASS range.endContainer is containerElement
PASS range.startOffset is 4
PASS range.endOffset is 5
PASS range.toString() is "3-begin 45 3-end"
range.setStart(containerElement, 5); range.setEnd(containerElement, 5); range.insertNode(insertedElement);
PASS range.startContainer is containerElement
PASS range.endContainer is containerElement
PASS range.startOffset is 4
PASS range.endOffset is 5
PASS range.toString() is "3-begin 45 3-end"
insertedElement = document.getElementById("div6");
range.setStart(containerElement, 1); range.setEnd(containerElement, 1); range.insertNode(insertedElement);
PASS range.startContainer is containerElement
PASS range.endContainer is containerElement
PASS range.startOffset is 1
PASS range.endOffset is 2
PASS range.toString() is "6"
insertedElement = document.getElementById("div3");
range.setStart(containerElement, 4); range.setEnd(containerElement, 4); range.insertNode(insertedElement);
PASS range.startContainer is containerElement
PASS range.endContainer is containerElement
PASS range.startOffset is 4
PASS range.endOffset is 5
PASS range.toString() is "3-begin 45 3-end"
PASS documentFragment = range.extractContents(); range.startContainer is containerElement
PASS range.startOffset is 4
PASS range.endContainer is containerElement
PASS range.endOffset is 4
PASS range.insertNode(documentFragment); range.toString() is "3-begin 45 3-end"
PASS successfullyParsed is true

TEST COMPLETE
